requirements.txt
openai-whisper==20240930
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==15.0.1
torch==2.7.0
torchaudio==2.7.0
numpy==1.26.4
scipy==1.12.0